Attach event handler to the document, because async
Sometimes, the code attaching the event handler fires before the form
is injected into the DOM. When this happens, an event handler targeting
the injected form doesn't get attached.

Instead, we attach the event handler to the document, because the event from the
form will bubble up to that layer.

This also targets the submit button being clicked, instead of the form being
submitted. This is because the form submit event is `preventDefault`ed in the
ConvertKit javascript.