Don't remove ck_subscriber_id from URL until we set a cookie
Previously, `ck_subscriber_id` was remvoed from the URL immediately on fetching
the query var. This worked fine in most cases, but if the AJAX call to the
plugin failed for any reason, then the cookie with `ck_subscriber_id` would never
be set, and the subscriber ID would be lost (until the subscriber clicked on
another email link with their subscriber ID included).

Now, we only remove the URL parameter if the AJAX call is successful and we
have set the cookie.