Feature/release process (#155)
Adds some tooling for building releases.

`.distignore` is used by `wp dist-archive` to build a zip of the plugin while leaving out files unwanted in the publicly distributed version.

`bin/package.sh` is a WIP script to build a public-ready release based on a GitHub release. This will be expanded in the future.