Handle cases where a category's form ID is 'default'
Before this, any category that had been edited (even with no changes) and saved
would result in any posts in that category never showing the global default form,
even if the post and category were both set to show the global default form.

This is because categories (which have been edited and saved) which have the default
form set have a `$form_id` of `'default'` set, but our code was only fetching
the global default form if the category's `$form_id` was `0`.

Closes #182