Update url library to work with PHP 7.3
Because PHP 7.3 moves from using PCRE to PCRE2 for regular expressions [1], some uses of `preg_match()` are doomed to failure on PHP 7.3.

The library we were using to convert relative to absolute URLs has not been updated for PHP 7.3.

I forked the library, and made the necessary change [2] so that it will work with PHP 7.3.

This commit removes the old version of the library, and instead loads our new fork.

[1] https://ayesh.me/Upgrade-PHP-7.3#pcre2
[2] https://github.com/ConvertKit/relative-to-absolute-url/commit/66426fa7afee86f1e6ff4f103e461aa6ec952c9b