Only return false from attempting to update stored content if failure
Prior to this change, we were directly using WordPress's built-in update_option()
function to store forms, tags, and landing pages. The problem with this is that
if any emoji were included in the content we attempted to store, the update_option()
call would fail if the site's database did not support the utf8mb4 character set,
and would return false.

However, it would also return false if there was simply no change between the prior
value and the new value! This is absurd.

Related: https://core.trac.wordpress.org/ticket/41099

Because of this, our attempted check for utf8mb4 problems was naively triggering in
this scenario:

1) Any of forms, tags, or LPs were unchanged when an update was attempted, AND
2) The site's database did not support the utf8mb4 character set, AND
3) No emoji characters were presnt in the content we attempted to save

In the above scenario, customers would get a false positive warning, which resulted
in an increase in support tickets related to this issue.

With this change, we first attempt the update_option() call. If, and only if, it
returns false, we then check to see if what we just attempted to save matches
what was already stored (by performing a get_option() call). If it does match,
then our wrapper function returns true, so our warning does not fire.

Closes #193