Don't show error on tools tab on first visit with logging on
Closes #196

Prior to this change, if a user visited the tools tab with logging turned on
on the general settings tab, AND had PHP error display turned on in e.g. wp-config.php,
then a PHP warning would be shown because our log.txt file would not exist, yet
we were trying to get the contents of it.

This change ensures that we only get the contents of the file if it already exists;
otherwise we show an empty string in the log text area.

This change also includes a Codeception test to cover this scenario 🔥🔥🔥