Save empty value for default form if API connection missing
Prior to this, if a site was connected to a ConvertKit account, then had its
API connection info removed, the previously set default form ID would still be
saved in wp_options.

With this change, if API connection info is removed, an empty value is saved
for the default form.