Only try to fetch a form for appending if ID is a non-array
Closes #204

Prior to this, therew was an edge case where if a site had the plugin installed
and activated without API credentials saved, then a situation could occur in
which the saved default form ID was an array with an error message.

If this happened, then a PHP notice & warning would be thrown when viewing a page
or post that would trigger a form being appended.

This change skips trying to append a form if the saved form ID value is an array.